+++ frontend/src/components/lobby/CreateRoomModal.tsx
@@ -5,7 +5,7 @@ import { useSocketStore } from "../../store/useSocket";
 import { GameRoom } from "../../store/useLobby";
 import { UserInfo, useRoomStore } from "../../store/useRoom";
 import { useNavigate } from "react-router";
-import { toast } from "../common/Toast";
+import { toast } from "../common/Toast/Toast";
 
 interface CreateRoomModalProps {
   closeModal: () => void;
